backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-migrate==4.0.7
SQLAlchemy==2.0.35
pymysql==1.1.1
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3
